+++ build_runner/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.7.6-dev
+
+- Update to package:build version `0.12.0`.
+- Removed the `DigestAssetReader` interface, the `digest` method has now moved
+  to the core `AssetReader` interface. We are treating this as a non-breaking
+  change because there are no known users of this interface.
+
 ## 0.7.5+1
 
 - Bug fix for using the `--output` flag when you have no `test` directory.